Split models into individual files
